Split up dists.ad.js.
