+++ .gitignore
@@ -8,10 +8,12 @@ npm-debug.log
 bundle/
 docs/_build
 docs/_templates
-src/dists.js
 src/dists.wppl
 src/inference/enumerate.js
 src/inference/elbo.js
 src/inference/eubo.js
 src/inference/dream/gradients.js
 src/aggregation/ScoreAggregator.js
+src/dists/*.js
+!/src/dists/*.ad.js
+src/math/numeric.js
