+++ scripts/distHeader
@@ -5,7 +5,7 @@
 var fs = require('fs');
 var path = require('path');
 var _ = require('lodash');
-var dists = require('../src/dists.ad.js');
+var dists = require('../src/dists');
 
 var helpers = {
   downcaseInitial: function(name) {
